fix check_trustlines test
