+++ polaris/polaris/tests/processes/test_check_trustlines.py
@@ -7,42 +7,34 @@ from polaris.tests.conftest import STELLAR_ACCOUNT_1
 from polaris.management.commands.check_trustlines import Command as CheckTrustlinesCMD
 
 
+account_json = {
+    "id": 1,
+    "sequence": 1,
+    "balances": [{"asset_code": "USD", "asset_issuer": USD_ISSUER_ACCOUNT,}],
+    "thresholds": {"low_threshold": 1, "med_threshold": 1, "high_threshold": 1,},
+    "signers": [{"key": STELLAR_ACCOUNT_1, "weight": 1}],
+}
 mock_server = Mock(
     accounts=Mock(
         return_value=Mock(
-            account_id=Mock(
-                return_value=Mock(
-                    call=Mock(
-                        return_value={
-                            "id": 1,
-                            "sequence": 1,
-                            "balances": [
-                                {
-                                    "asset_code": "USD",
-                                    "asset_issuer": USD_ISSUER_ACCOUNT,
-                                }
-                            ],
-                            "thresholds": {
-                                "low_threshold": 1,
-                                "med_threshold": 1,
-                                "high_threshold": 1,
-                            },
-                            "signers": [{"key": STELLAR_ACCOUNT_1, "weight": 1}],
-                        }
-                    )
-                )
-            )
+            account_id=Mock(return_value=Mock(call=Mock(return_value=account_json)))
         )
     )
 )
 
 
 @pytest.mark.django_db
-@patch("polaris.management.commands.check_trustlines.PendingDeposits")
+@patch("polaris.management.commands.check_trustlines.PendingDeposits.submit")
 @patch(
     "polaris.management.commands.check_trustlines.settings.HORIZON_SERVER", mock_server
 )
-def test_deposit_check_trustlines_success(acc1_usd_deposit_transaction_factory):
+@patch(
+    "polaris.management.commands.check_trustlines.MultiSigTransactions.requires_multisig",
+    Mock(return_value=False),
+)
+def test_deposit_check_trustlines_success(
+    mock_submit, acc1_usd_deposit_transaction_factory
+):
     """
     Creates a transaction with status `pending_trust` and checks that
     `check_trustlines` changes its status to `completed`. All the necessary
@@ -51,6 +43,8 @@ def test_deposit_check_trustlines_success(acc1_usd_deposit_transaction_factory):
     deposit = acc1_usd_deposit_transaction_factory()
     deposit.status = Transaction.STATUS.pending_trust
     deposit.save()
-    deposit.refresh_from_db = Mock()
+    mock_submit.return_value = True
+
     CheckTrustlinesCMD.check_trustlines()
-    assert deposit.refresh_from_db.was_called
+
+    mock_submit.assert_called_once_with(deposit)
